+++ genki_wave/quaternions.py
@@ -1,6 +1,47 @@
-from typing import Union
+import math
+from dataclasses import dataclass
 
-from genki_wave.data.organization import ButtonEvent, DataPackage, Point3d, Quaternion
+from genki_wave.data.organization_elements import Point3d
+
+
+@dataclass(frozen=True)
+class Quaternion:
+    w: float
+    x: float
+    y: float
+    z: float
+
+    @classmethod
+    def from_point3d(cls, p: Point3d) -> "Quaternion":
+        return cls(0, p.x, p.y, p.z)
+
+    def to_point3d(self):
+        return Point3d(self.x, self.y, self.z)
+
+    def __mul__(self, other):
+        w1, x1, y1, z1 = self.w, self.x, self.y, self.z
+        w2, x2, y2, z2 = other.w, other.x, other.y, other.z
+
+        w = w1 * w2 - x1 * x2 - y1 * y2 - z1 * z2
+        x = w1 * x2 + x1 * w2 + y1 * z2 - z1 * y2
+        y = w1 * y2 + y1 * w2 + z1 * x2 - x1 * z2
+        z = w1 * z2 + z1 * w2 + x1 * y2 - y1 * x2
+        return Quaternion(w, x, y, z)
+
+    def conjugate(self):
+        return Quaternion(self.w, -self.x, -self.y, -self.z)
+
+    def normalize(self):
+        norm = math.sqrt(sum([el**2 for el in [self.w, self.x, self.y, self.z]]))
+        return Quaternion(self.w / norm, self.x / norm, self.y / norm, self.z / norm)
+
+
+def rotate_vector(p: Point3d, q: Quaternion) -> Point3d:
+    """Rotate point p by quaternion q"""
+    q = q.normalize()
+    p = Quaternion.from_point3d(p)
+    p_rot = q * p * q.conjugate()
+    return p_rot.to_point3d()
 
 
 def quat_mult(q1: Quaternion, q2: Quaternion) -> Quaternion:
@@ -17,15 +58,12 @@ def quat_mult(q1: Quaternion, q2: Quaternion) -> Quaternion:
 def quat_conj(q):
     w, x, y, z = q
     return w, -x, -y, -z
+#
+#
+# def rotate_vector(x: Point3d, q: Quaternion):
+#     q_conj = quat_conj(q)
+#     x_quat = [0, x[0], x[1], x[2]]
+#     out = quat_mult(quat_mult(q, x_quat), q_conj)
+#     return out[1:]
 
 
-def rotate_vector(x: Point3d, q: Quaternion):
-    q_conj = quat_conj(q)
-    x_quat = [0, x[0], x[1], x[2]]
-    out = quat_mult(quat_mult(q, x_quat), q_conj)
-    return out[1:]
-
-
-def acc_global(data: Union[DataPackage, ButtonEvent]):
-    data
-    pass
