+++ tests/constants.py
@@ -4,9 +4,9 @@ from genki_wave.data.organization import (
     ButtonId,
     DataPackage,
     Euler3d,
-    Point3d,
     Quaternion,
 )
+from genki_wave.data.organization_elements import Point3d
 
 SERIAL_DATA = [
     b"\x012\x807\x87b\xbfrL^\xbe\x12\t>\xbd\x8e\x8f\xcf\xbe7\x87b\xbfrL^\xbe\x12\t>\xbd\x8e\x8f\xcf\xbeQ\x15\xe0>3"
